chore: define createKeyLock and refactor Realize Agent about lock
